+++ test/types.ts
@@ -1,6 +1,8 @@
 import {
   evaluate,
-  unaryTest
+  unaryTest,
+  compileExpression,
+  compileUnaryTests
 } from '..';
 
 import { expect } from 'chai';
@@ -31,6 +33,46 @@ describe('types', function() {
 
       // when
       const { value, warnings } = unaryTest('[10, 20]', {
+        '?': 5
+      });
+
+      // then
+      expect(value).to.be.false;
+      expect(warnings).to.be.an('array');
+    });
+
+  });
+
+
+  describe('compileExpression', function() {
+
+    it('should evaluate compiled artifact', function() {
+
+      // when
+      const compiled = compileExpression('hello', {
+        hello: 'HELLO'
+      });
+
+      const { value, warnings } = compiled.evaluate({
+        hello: 'HELLO'
+      });
+
+      // then
+      expect(value).to.eql('HELLO');
+      expect(warnings).to.eql([]);
+    });
+
+  });
+
+
+  describe('compileUnaryTests', function() {
+
+    it('should test compiled artifact', function() {
+
+      // when
+      const compiled = compileUnaryTests('[10, 20]');
+
+      const { value, warnings } = compiled.unaryTest({
         '?': 5
       });